--- composeApp/src/commonMain/kotlin/com/aman/core/domain/Error.kt
@@ -1,3 +1,0 @@
-package com.aman.core.domain
-
-interface Error
