+++ composeApp/src/commonMain/kotlin/com/aman/currencyconverter/core/domain/DataError.kt
@@ -1,0 +1,17 @@
+package com.aman.currencyconverter.core.domain
+
+sealed interface DataError: Error {
+    enum class Remote: DataError {
+        REQUEST_TIMEOUT,
+        TOO_MANY_REQUESTS,
+        NO_INTERNET,
+        SERVER,
+        SERIALIZATION,
+        UNKNOWN
+    }
+
+    enum class Local: DataError {
+        DISK_FULL,
+        UNKNOWN
+    }
+}
