+++ composeApp/src/commonMain/kotlin/com/aman/currencyconverter/core/domain/Error.kt
@@ -1,0 +1,3 @@
+package com.aman.currencyconverter.core.domain
+
+interface Error
