--- composeApp/src/commonMain/kotlin/com/aman/currencyconverter/presentation/model/CurrencyConverterAction.kt
@@ -1,9 +1,0 @@
-package com.aman.currencyconverter.presentation.model
-
-sealed class CurrencyConverterAction {
-    data class FromCurrencyChanged(val currency: String) : CurrencyConverterAction()
-    data class ToCurrencyChanged(val currency: String) : CurrencyConverterAction()
-    data class AmountChanged(val amount: String) : CurrencyConverterAction()
-    data object Convert : CurrencyConverterAction()
-    data object LoadCurrencies : CurrencyConverterAction()
-}
